Move change file validation code into ChangeFiles class
