+++ apps/rush/src/utilities/test/PublishUtilities.test.ts
@@ -390,24 +390,3 @@ describe('isRangeDependency', () => {
     expect(PublishUtilities.isRangeDependency('~1.0.0')).is.false;
   });
 });
-
-describe('findMissingChangedPackages', () => {
-  it('finds the missing package.', () => {
-    const changeFile: string = path.join(__dirname, 'verifyChanges', 'changes.json');
-    const changedPackages: string[] = ['a', 'b', 'c'];
-    expect(PublishUtilities.findMissingChangedPackages(changeFile, changedPackages)).to.contain(
-      'c',
-      'c should be missing');
-    expect(PublishUtilities.findMissingChangedPackages(changeFile, changedPackages)).to.have.lengthOf(
-      1,
-      'only c is missing');
-  });
-
-  it('finds nothing when no missing packages', () => {
-    const changeFile: string = path.join(__dirname, 'verifyChanges', 'changes.json');
-    const changedPackages: string[] = ['a'];
-    expect(PublishUtilities.findMissingChangedPackages(changeFile, changedPackages)).to.have.lengthOf(
-      0,
-      'nothing is missing');
-  });
-});
